# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Helix
  version: 0.1.0
  description: |
    Local DSPy compile/eval job runner. Single-port (127.0.0.1:7000),
    Postgres-durable queue, blob-stored artifacts, Langfuse traces via
    same-origin auto-SSO trampoline.

    This spec is the source of truth. Server (FastAPI), UI (Next.js),
    and CLI (Python) all consume generated types from it; no hand-
    written models or request bodies.
servers:
  - url: http://127.0.0.1:7000
    description: Local Helix instance (behind Caddy)
...
paths:
  /api/jobs/compile:
    post:
      tags: [jobs]
      operationId: submitCompile
      summary: Submit a compile job
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CompileSubmission'
            encoding:
              metadata:
                contentType: application/json
              bundle:
                contentType: application/gzip
      responses:
        '201':
          description: Job(s) created
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/JobSubmissionResult' }
        '400': { $ref: '#/components/responses/Error' }
        '409':
          description: Worker fleet has divergent baked_sha or stack not ready
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/jobs/import-compile:
    post:
      tags: [jobs]
      operationId: importCompile
      summary: Import a legacy local results directory as an ad-hoc compile job
      description: |
        Accepts a tarball of a results-dir tree (as produced by the old
        launcher). Creates a `succeeded` compile job row with
        baked_sha=null, emitted_run_number parsed from the directory
        name, and uploads every file (except evals/** and helix/**)
        as an artifact. Idempotent on (program_version, emitted_run_number).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImportCompileSubmission'
            encoding:
              metadata: { contentType: application/json }
              bundle:   { contentType: application/gzip }
      responses:
        '200':
          description: Existing job returned (idempotent hit)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobSubmissionResult' }
        '201':
          description: New imported job created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobSubmissionResult' }
        '400': { $ref: '#/components/responses/Error' }
...
  /api/jobs/{job_id}:
    get:
      tags: [jobs]
      operationId: getJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '404': { $ref: '#/components/responses/Error' }
...
  /api/jobs/{job_id}/artifacts:
    get:
      tags: [artifacts]
      operationId: listArtifacts
      parameters:
        - $ref: '#/components/parameters/JobId'
        - { name: kind,   in: query, schema: { type: string } }
        - { name: prefix, in: query, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Artifact' }
  /api/jobs/{job_id}/artifacts/{artifact_id}:
    get:
      tags: [artifacts]
      operationId: getArtifact
      parameters:
        - $ref: '#/components/parameters/JobId'
        - { name: artifact_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Raw artifact stream
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '404': { $ref: '#/components/responses/Error' }
  /api/jobs/{job_id}/artifacts.tar.gz:
    get:
      tags: [artifacts]
      operationId: downloadArtifactsTar
      summary: Bulk tar.gz of all (or prefix-filtered) artifacts at their relative_path
      parameters:
        - $ref: '#/components/parameters/JobId'
        - { name: prefix, in: query, schema: { type: string } }
      responses:
        '200':
          description: tar.gz stream
          content:
            application/gzip:
              schema: { type: string, format: binary }
...
components:
  parameters:
    JobId:
      name: job_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:    { type: string, example: "validation_error" }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
    JobType:
      type: string
      enum: [compile, eval]
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled, blocked]
...
    Job:
      type: object
      required:
        - id
        - type
        - status
        - run_label
        - attempt
        - created_at
      properties:
        id:                  { type: string, format: uuid }
        type:                { $ref: '#/components/schemas/JobType' }
        status:              { $ref: '#/components/schemas/JobStatus' }
        repo_id:             { type: string,  nullable: true, description: "trusted-local multi-repo key" }
        snapshot_id:         { type: string,  format: uuid, nullable: true, description: "content-addressed snapshot the job runs against" }
        blocked_reason:      { type: string,  nullable: true, description: "why a 'blocked' job could not run (e.g. snapshot unavailable)" }
        program:             { type: string,  nullable: true }
        version:             { type: string,  nullable: true }
        dataset:             { type: string,  nullable: true }
        split:               { type: string,  nullable: true }
        parent_job_id:       { type: string,  format: uuid, nullable: true }
        config_path:         { type: string,  nullable: true, description: "repo-relative path of the submitted config" }
        baked_sha:           { type: string,  nullable: true, description: "legacy; superseded by snapshot_id" }
        run_label:           { type: string }
        attempt:             { type: integer }
        worker_id:           { type: string,  nullable: true }
        lease_expires_at:    { type: string,  format: date-time, nullable: true }
        emitted_run_number:  { type: integer, nullable: true }
        export_run_number:   { type: integer, nullable: true }
        created_at:          { type: string,  format: date-time }
        started_at:          { type: string,  format: date-time, nullable: true }
        ended_at:            { type: string,  format: date-time, nullable: true }
        exit_code:           { type: integer, nullable: true }
        summary:
          type: object
          additionalProperties: true
          nullable: true
          description: "Final metrics parsed from the worker's status line (acc, cost, tokens, latency)."
        ui_url:              { type: string, description: "Absolute URL to this job in the UI" }
        traces_url:          { type: string, description: "Absolute URL to the SSO trampoline that lands on this job's traces" }
    JobSubmissionResult:
      type: object
      required: [job_id, run_label, ui_url, traces_url]
      properties:
        job_id:     { type: string, format: uuid }
        run_label:  { type: string }
        ui_url:     { type: string }
        traces_url: { type: string }
    CompileSubmissionMetadata:
      type: object
      required: [repo_id, configs]
      properties:
        repo_id:                { type: string, description: "trusted-local multi-repo key" }
        snapshot_id:            { type: string, format: uuid, nullable: true, description: "published snapshot the job runs against" }
        snapshot_digest:        { type: string, nullable: true }
        helix_runtime_version:  { type: string, nullable: true, description: "consumer's required helix-runtime version spec" }
        baked_sha:              { type: string, nullable: true, description: "legacy; superseded by snapshot_id" }
        program:                { type: string, nullable: true, description: "override; otherwise inferred from config path" }
        version:                { type: string, nullable: true }
        configs:
          type: array
          minItems: 1
          items:
            type: object
            required: [config_path, dataset, split]
            properties:
              config_path: { type: string, description: "repo-relative path under <overlay-root>/<p>/<v>/" }
              dataset:     { type: string }
              split:       { type: string }
              auto_eval_config_path:
                type: string
                nullable: true
                description: "If set, auto-chain an eval against this compile on success."
        overlay_files:
          type: array
          items: { type: string, description: "repo-relative paths in the uploaded bundle" }
    CompileSubmission:
      type: object
      required: [metadata, bundle]
      properties:
        metadata: { $ref: '#/components/schemas/CompileSubmissionMetadata' }
        bundle:
          type: string
          format: binary
          description: "tar.gz overlay; may be empty if config is identical to baked snapshot"
...
    ImportCompileMetadata:
      type: object
      required: [repo_id, program, version, dataset, split, emitted_run_number, results_dir_basename]
      properties:
        repo_id:               { type: string, description: "trusted-local multi-repo key" }
        program:               { type: string }
        version:               { type: string }
        dataset:               { type: string }
        split:                 { type: string }
        emitted_run_number:    { type: integer }
        results_dir_basename:  { type: string }
        compile_config_path:   { type: string, nullable: true }
    ImportCompileSubmission:
      type: object
      required: [metadata, bundle]
      properties:
        metadata: { $ref: '#/components/schemas/ImportCompileMetadata' }
        bundle:
          type: string
          format: binary
          description: "tar.gz of the entire local results-dir (excluding evals/** and helix/**)"
    Artifact:
      type: object
      required: [id, job_id, relative_path, kind, size_bytes, sha256, attempt, created_at]
      properties:
        id:             { type: string, format: uuid }
        job_id:         { type: string, format: uuid }
        relative_path:  { type: string }
        kind:           { type: string }
        size_bytes:     { type: integer, format: int64 }
        sha256:         { type: string, minLength: 64, maxLength: 64 }
        mime:           { type: string, nullable: true }
        attempt:        { type: integer }
        created_at:     { type: string, format: date-time }
...